--- src/translation/translationObject.ts
@@ -1,7 +1,0 @@
-import en from './locales/en/translations.json';
-import ru from './locales/ru/translations.json';
-
-export const translationObject = {
-  en,
-  ru,
-};
